+++ src/msgpack-autosplit.h
@@ -28,7 +28,7 @@ typedef struct AppContext_ {
     time_t              logfile_last_rotation;
     time_t              logfile_rotate_after;
     unsigned int        logfile_max_files;
-    size_t              logfile_max_space;
+    off_t               logfile_max_space;
     unsigned int        logfile_seq;
     LogCompression      log_compression;
     _Bool               logfile_enabled;
